Move formatDuration to common package
Remove parseDuration as it isn't used.
Align cell.time tooltip with prRevision.
